utils/export-articles-to-lipu-tenpo/pyproject.toml
[project]
name = "export-articles-to-lipu-tenpo"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = ["pyyaml>=6.0.3"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
